Completed All buttons and light functionalities
Added 2 LED.  1 LED  will light up when the hardware receives information from the App.  The other LED will turn on if there is a problem sending updates to the API.  There are now 3 buttons.  The first button locks and unlocks the driver's door.  The second button turns on the driver's seat warmer, and the third button turns on the front defrost.  Finally a timer was created.  This timer will turn off both defrost after 1 minute.